refactor: StatusPartMaker での分割がいい感じだったので、すべてのパートをMakerクラス化。
